requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
cramjam==2.10.0
distro==1.9.0
fastparquet==2024.11.0
fsspec==2025.3.2
google-auth==2.39.0
google-genai==1.13.0
google_search_results==2.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
numpy==2.2.5
openai==1.76.2
packaging==25.0
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
serpapi==0.1.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
